requirements.txt
flask
pandas
openpyxl
Office365-REST-Python-Client
gunicorn